Create resourceAllocation.c
Ques. Consider that a system has P resources of the same type. These resources are shared by Q
processes from time to time. All processes request and release the resources one at a time. Generate a
solution to demonstrate that, the system is in a safe state when the following conditions are satisfied.
Conditions:
1. The maximum resource need of each process is between 1 and P.
2. Summation of all maximum needs is less than P+Q