largest burst time scheduler
Design a scheduler following a non-preemptive scheduling approach to schedule the
processes that arrive at different units and having burst time double the arrival time. The scheduler
selects the process with the largest burst time from the queue for the execution. The process is not being
preempted until it finishes its service time. Compute the average waiting time and average
turnaround time.